Use MIT license
CC0-1.0 is not recognised by the Open Source Initiative (https://opensource.org) and was only there from the boilerplate code that this repo was based on. MIT is popular and seems like the most similar/permissive officially recognised license.

I believe any releases prior to this commit will still fall under CC0-1.0 if people have a problem with MIT.